web/app/requirements.txt
absl-py==0.9.0
alembic==1.4.1
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==17.4.0
Automat==0.6.0
blinker==1.4
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloud-init==19.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coverage==5.0.3
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
distro-info===0.18ubuntu0.18.04.1
dlib==19.19.0
dominate==2.5.1
en-core-web-sm==2.2.5
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
httpie==2.0.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
importlib-metadata==1.6.0
incremental==16.10.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
murmurhash==1.0.2
netifaces==0.10.4
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==20.3
PAM==0.4.2
pandas==1.0.1
Pillow==7.1.2
pip-review==1.0
plac==1.1.3
preshed==3.0.2
protobuf==3.11.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pylint==2.4.4
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-docx==0.8.10
python-editor==1.0.4
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
scipy==1.4.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.14.0
spacy==2.2.4
SQLAlchemy==1.3.15
srsly==1.0.2
ssh-import-id==5.7
systemd-python==234
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
thinc==7.4.0
torch==1.4.0
torchdiffeq==0.0.1
tqdm==4.45.0
Twisted==17.9.0
typed-ast==1.4.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
visitor==0.1.3
wasabi==0.6.0
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.3.2
